@@ -8,24 +8,25 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "fideloper/proxy": "^4.3",
-        "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.5",
-        "laravel/framework": "^7.5",
-        "laravel/tinker": "^2.4",
-        "laravel/ui": "^2.0",
-        "laravel/passport": "^8.4",
-        "doctrine/dbal": "^2.10",
-        "yajra/laravel-datatables-oracle": "^9.9",
-        "bugsnag/bugsnag-laravel": "^2.18"
+        "php": "^7.3|^8.0",
+        "bugsnag/bugsnag-laravel": "^2.21",
+        "doctrine/dbal": "^3.0",
+        "fideloper/proxy": "^4.4",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.2.0",
+        "laravel/framework": "^8.18",
+        "laravel/passport": "^10.1",
+        "laravel/tinker": "^2.5",
+        "laravel/ui": "^3.1",
+        "yajra/laravel-datatables-oracle": "^9.15"
     },
     "require-dev": {
-        "facade/ignition": "^2.0",
-        "fzaninotto/faker": "^1.9",
-        "mockery/mockery": "^1.3",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^9.1"
+        "facade/ignition": "^2.5.3",
+        "fakerphp/faker": "^1.12.1",
+        "laravel/sail": "^0.0.9",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.1",
+        "phpunit/phpunit": "^9.5.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -39,12 +40,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
