@@ -2,9 +2,9 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Channel;
+use App\Models\Video;
 use Illuminate\Http\Request;
-use App\Video;
-use App\Channel;
 
 class HomePageController extends Controller
 {
