@@ -1,15 +1,15 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Role extends Model
+class Permission extends Model
 {
     use SoftDeletes;
 
-    public $table = 'roles';
+    public $table = 'permissions';
 
     protected $dates = [
         'created_at',
@@ -23,10 +23,4 @@ class Role extends Model
         'updated_at',
         'deleted_at',
     ];
-
-    public function permissions()
-    {
-        return $this->belongsToMany(Permission::class);
-
-    }
 }
