@@ -1,15 +1,15 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Permission extends Model
+class Role extends Model
 {
     use SoftDeletes;
 
-    public $table = 'permissions';
+    public $table = 'roles';
 
     protected $dates = [
         'created_at',
@@ -23,4 +23,10 @@ class Permission extends Model
         'updated_at',
         'deleted_at',
     ];
+
+    public function permissions()
+    {
+        return $this->belongsToMany(Permission::class);
+
+    }
 }
